feat&docs: add animSwitch story move switches' stories into settings section create useTryQueryClient hook
